requirements.txt
Flask==2.3.3
qrcode==7.3.1
PyPDF2==3.0.1
gunicorn==20.1.0
boto3==1.28.7     # AWS SDK for Python, used to interact with Amazon S3
python-dotenv==1.0.0  # Library for loading environment variables from .env files
werkzeug==2.3.7 